rename: Initramfs -> Cpiofs
