+++ kernel/src/fs/drivers/cpio/tests/mkfs/initramfs/file1.txt
