+++ kernel/src/fs/drivers/mod.rs
@@ -1,1 +1,1 @@
-pub mod initramfs;
+pub mod cpio;
